Sizes of cursors and pages are integers in REST. (#3938)
This change addresses the problem of uint64 not being representable in
javascript, leading to a swagger definition of two numbers related to
cursors, namely `size` and `total_records`, having string type rather
than integer. Using a uint32 generates the desired type and is
sufficient for paging purposes.

This is a breaking change, which is why this commit contains changes
to CLI as well. For the record, the feature is currently under feature
flag.

Fixes #3934

+++ internal/controlplane/handlers_evalstatus.go
@@ -36,11 +36,11 @@ import (
 )
 
 const (
-	defaultPageSize uint64 = 25
+	defaultPageSize uint32 = 25
 	// Maximum page size has a conservative value at the moment,
 	// we can raise it once we have more insight on its
 	// performance impact.
-	maxPageSize uint64 = 25
+	maxPageSize uint32 = 25
 )
 
 // ListEvaluationHistory lists current and past evaluation results for
@@ -193,7 +193,7 @@ func fromEvaluationHistoryRow(
 	return res, nil
 }
 
-func makeCursor(cursor []byte, size uint64) *minderv1.Cursor {
+func makeCursor(cursor []byte, size uint32) *minderv1.Cursor {
 	return &minderv1.Cursor{
 		Cursor: base64.StdEncoding.EncodeToString(cursor),
 		Size:   size,
